More vocabularies added to the annotation

+++ omop_cdm/generate_code_lookup_json.py
@@ -1,6 +1,7 @@
 """
-    The goal is to build dictionaries that can be used to search for codes in
-    sources and map to elements in the CDM vocabulary
+    The goal is to build JSON dictionaries that can be used to search for codes in
+    sources and map to elements in the CDM vocabulary.
+
 """
 
 import csv
@@ -16,6 +17,7 @@ def main(source_vocabulary_directory, output_json_directory=None, delimiter="\t"
     concept_csv = os.path.join(source_vocabulary_directory, "CONCEPT.csv")
     vocabularies = []
 
+    # Determine which vocabularies are in the concept file
     with open(concept_csv, "rb") as f:
         dict_reader = csv.DictReader(f, delimiter=delimiter)
         for row_dict in dict_reader:
@@ -23,6 +25,8 @@ def main(source_vocabulary_directory, output_json_directory=None, delimiter="\t"
             if vocabulary_id not in vocabularies:
                 vocabularies += [vocabulary_id]
 
+    # Generate first pass of converting concepts into a JSON lookup file
+    # Generate one for concept_name and one for concept_code
     for vocabulary in vocabularies:
         fields_to_key_on = ["CONCEPT_CODE", "CONCEPT_NAME"]
 
@@ -36,7 +40,10 @@ def main(source_vocabulary_directory, output_json_directory=None, delimiter="\t"
 
     concept_relationship_csv = os.path.join(source_vocabulary_directory, "CONCEPT_RELATIONSHIP.CSV")
     concept_relationship_json = os.path.join(output_json_directory, "concept_relationship.json")
-    csv_file_name_to_keyed_json(concept_relationship_csv, concept_relationship_json, "CONCEPT_ID_1", ("RELATIONSHIP_ID", "Maps to"))
+    # Build a master dict
+    if not(os.path.exists(concept_relationship_json)):
+        print("Generating '%s'" % concept_relationship_json)
+        csv_file_name_to_keyed_json(concept_relationship_csv, concept_relationship_json, "CONCEPT_ID_1", ("RELATIONSHIP_ID", "Maps to"))
 
     with open(concept_csv, "rb") as f:
         dict_reader = csv.DictReader(f, delimiter=delimiter)
@@ -45,49 +52,52 @@ def main(source_vocabulary_directory, output_json_directory=None, delimiter="\t"
             concept_dict_vocabulary[row_dict["CONCEPT_ID"]] = row_dict["VOCABULARY_ID"]
 
         global_concept_json = os.path.join(output_json_directory, "global_concept_vocabulary.json")
-        print("Writing '%s" % global_concept_json)
+        print("Generating '%s" % global_concept_json)
         with open(global_concept_json, "w") as fw:
             json.dump(concept_dict_vocabulary, fw, sort_keys=True, indent=4, separators=(',', ': '))
 
-    vocabularies_with_maps = ["ICD9CM"]
+    vocabularies_with_maps = ["ICD9CM", "ICD9Proc", "ICD10CM", "ICD10PCS", "Multum", "LOINC", "CPT4"]
     for vocabulary_id in vocabularies_with_maps:
-
+        print("Annotating '%s'" % vocabulary_id)
         vocabulary_json = os.path.join(output_json_directory, "CONCEPT_CODE_" + vocabulary_id + ".json")
-        with open(vocabulary_json, "r") as f:
-            vocabulary_dict = json.load(f)
-
-        with open(concept_relationship_json, "r") as f:
-            concept_rel_dict = json.load(f)
-
-        for concept_code in vocabulary_dict:
-            concept_dict = vocabulary_dict[concept_code]
-            concept_id = concept_dict["CONCEPT_ID"]
-            if concept_id in concept_rel_dict:
-                try:
-                    mapped_concept_id = concept_rel_dict[concept_id]["CONCEPT_ID_2"]
-                except TypeError:
-                    multiple_concepts = concept_rel_dict[concept_id]
-                    multiple_concepts.sort(key=lambda x: x["VALID_END_DATE"], reverse=True)
-                    mapped_concept_id = multiple_concepts[0]["CONCEPT_ID_2"]
-
-                concept_dict["MAPPED_CONCEPT_ID"] = mapped_concept_id
-                if mapped_concept_id in concept_dict_vocabulary:
-                    concept_dict["MAPPED_CONCEPT_VOCAB"] = concept_dict_vocabulary[mapped_concept_id]
-                else:
-                    concept_dict["MAPPED_CONCEPT_VOCAB"] = None
-
-            else:
-                concept_dict["MAPPED_CONCEPT_ID"] = None
 
         concept_with_parent_json = os.path.join(output_json_directory, vocabulary_id + "_with_parent.json")
-        with open(concept_with_parent_json, "w") as fw:
-            json.dump(vocabulary_dict, fw, sort_keys=True, indent=4, separators=(',', ': '))
+
+        if not os.path.exists(concept_with_parent_json):
+            with open(vocabulary_json, "r") as fj:
+                vocabulary_dict = json.load(fj)
+
+            with open(concept_relationship_json, "r") as fj:
+                concept_rel_dict = json.load(fj)
+
+            for concept_code in vocabulary_dict:
+                concept_dict = vocabulary_dict[concept_code]
+                concept_id = concept_dict["CONCEPT_ID"]
+                if concept_id in concept_rel_dict:
+                    try:
+                        mapped_concept_id = concept_rel_dict[concept_id]["CONCEPT_ID_2"]
+                    except TypeError:
+                        multiple_concepts = concept_rel_dict[concept_id]
+                        multiple_concepts.sort(key=lambda x: x["VALID_END_DATE"], reverse=True)
+                        mapped_concept_id = multiple_concepts[0]["CONCEPT_ID_2"]
+
+                    concept_dict["MAPPED_CONCEPT_ID"] = mapped_concept_id
+                    if mapped_concept_id in concept_dict_vocabulary:
+                        concept_dict["MAPPED_CONCEPT_VOCAB"] = concept_dict_vocabulary[mapped_concept_id]
+                    else:
+                        concept_dict["MAPPED_CONCEPT_VOCAB"] = None
+
+                else:
+                    concept_dict["MAPPED_CONCEPT_ID"] = None
+
+            with open(concept_with_parent_json, "w") as fw:
+                json.dump(vocabulary_dict, fw, sort_keys=True, indent=4, separators=(',', ': '))
 
 
 def csv_file_name_to_keyed_json(csv_file_name, json_file_name, field_to_key_on, filter_pair=None, delimiter="\t"):
     """Create a keyed JSON file"""
-    with open(csv_file_name, "rb") as f:
-        dict_reader = csv.DictReader(f, delimiter=delimiter)
+    with open(csv_file_name, "rb") as fd:
+        dict_reader = csv.DictReader(fd, delimiter=delimiter)
         result_dict = {}
 
         include_row = True
@@ -121,4 +131,6 @@ def csv_file_name_to_keyed_json(csv_file_name, json_file_name, field_to_key_on, 
 
 
 if __name__ == "__main__":
-    main("E:\\data\\vocab_download_v5_{680A7A12-692C-D9A0-351D-DB05A1E8A46D}")
+    with open("hi_config.json") as f:
+        cdm_config = json.load(f)
+    main(cdm_config["json_map_directory"])
